Move and rename `Admin\Product\Attributes\AttributesBlock` class to `Admin\ProductBlocksService` class.
